Change terminator to use separate config files
Similar to the i3 config. In terminator, configlaptop will be the
laptop config and configdesktop the desktop config. The keyboard
shortcut to open terminator in .config/i3/config will then point
towards the correct file. This was done because font sizes never
looked good on the vastly different screen sizes.